refactor:cleaned up frontend files and structure
