+++ frontend/components/Procurement/ContactsTable.js
@@ -9,7 +9,6 @@ export default function contacts() {
   const fetchData = () => {
     fetch(`http://localhost:8787/${keyword}`)
       .then((response) => response.json())
-      .then(alert(data))
       .then((data) => setData(data._embedded.supplierList))
       .catch((err) => console.error(err));
   };
@@ -19,16 +18,12 @@ export default function contacts() {
   }, []);
 
   const columns = [
-    { field: "supplier_id", headerName: "Supplier ID", width: 125, minWidth: 150, maxWidth: 200 },
+    { field: "supplier_id", headerName: "Supplier ID", width: 100, minWidth: 100, maxWidth: 200 },
+    { field: "id", headerName: "Contact ID", width: 100, minWidth: 100, maxWidth: 200 },
     { field: "name", headerName: "Name", width: 125, minWidth: 150, maxWidth: 200 },
     { field: "phone", headerName: "Phone", width: 125, minWidth: 150, maxWidth: 200 },
-    {
-      field: "email",
-      headerName: "Email",
-      width: 125, minWidth: 150, maxWidth: 200
-    },
-    { field: "position", headerName: "Position", width: 125, minWidth: 150, maxWidth: 200 },
-    { field: "id", headerName: "Contact ID", width: 125, minWidth: 150, maxWidth: 200 },
+    { field: "email", headerName: "Email", width: 200, minWidth: 200, maxWidth: 300},
+    { field: "position", headerName: "Position", width: 200, minWidth: 200, maxWidth: 300 },
   ];
 
   const rows = [];
